refactor: decouple default service name from schematize_service_name (#19030)
Separate default service-name resolution from the versioned span schema so settings no longer pull the full schema dependency graph into initialization.

(cherry picked from commit 727771442208e04aec8293befd6826a3dd7d4e45)
