+++ tests/contrib/aiopg/test.py
@@ -6,7 +6,7 @@ import pytest
 
 from ddtrace.contrib.internal.aiopg.patch import patch
 from ddtrace.contrib.internal.aiopg.patch import unpatch
-from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
 
 # project
 from ddtrace.trace import Pin
