+++ tests/contrib/boto/test.py
@@ -17,7 +17,7 @@ from moto import mock_sts
 from ddtrace.contrib.internal.boto.patch import patch
 from ddtrace.contrib.internal.boto.patch import unpatch
 from ddtrace.ext import http
-from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
 
 # project
 from ddtrace.trace import Pin
