+++ tests/contrib/grpc/test_grpc.py
@@ -13,7 +13,7 @@ from ddtrace.contrib.internal.grpc import constants
 from ddtrace.contrib.internal.grpc.patch import _unpatch_server
 from ddtrace.contrib.internal.grpc.patch import patch
 from ddtrace.contrib.internal.grpc.patch import unpatch
-from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
 from ddtrace.trace import Pin
 from tests.utils import TracerTestCase
 from tests.utils import snapshot
