+++ tests/contrib/mysqldb/test_mysqldb.py
@@ -4,7 +4,7 @@ import pytest
 
 from ddtrace.contrib.internal.mysqldb.patch import patch
 from ddtrace.contrib.internal.mysqldb.patch import unpatch
-from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
 from ddtrace.trace import Pin
 from tests.contrib import shared_tests
 from tests.opentracer.utils import init_tracer
