+++ tests/contrib/psycopg2/test_psycopg.py
@@ -9,7 +9,7 @@ from psycopg2 import extras
 
 from ddtrace.contrib.internal.psycopg.patch import patch
 from ddtrace.contrib.internal.psycopg.patch import unpatch
-from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
 from ddtrace.internal.utils.version import parse_version
 from ddtrace.trace import Pin
 from tests.contrib.config import POSTGRES_CONFIG
