+++ tests/contrib/pymysql/test_pymysql.py
@@ -3,7 +3,7 @@ import pymysql
 
 from ddtrace.contrib.internal.pymysql.patch import patch
 from ddtrace.contrib.internal.pymysql.patch import unpatch
-from ddtrace.internal.schema import DEFAULT_SPAN_SERVICE_NAME
+from ddtrace.internal.schema.default import DEFAULT_SPAN_SERVICE_NAME
 from ddtrace.trace import Pin
 from tests.contrib import shared_tests
 from tests.opentracer.utils import init_tracer
